requirements.txt
discord.py
apscheduler
flask
python-dotenv
loguru
pandas
pytz